refactor: move _async_update_data and _async_start_poll_cycle to PollingOperations
Move the two largest coordinator methods (~905 lines total) from main.py
to polling.py (PollingOperations mixin):

- _async_update_data: Core polling trigger method (~455 lines)
- _async_start_poll_cycle: Sequential device polling (~450 lines)

Also extract row sanitization helpers to helpers/cache.py:
- row_source_label: Determine source rank/label for reports
- sanitize_decoder_row: Normalize decoder row for HA attributes
- get_common_pb2: Lazy protobuf loader

main.py is now ~2,721 lines (down from ~8,342 originally).
